tg_bot/tg_bot.py
import logging

from telegram import Update
from telegram.ext import ApplicationBuilder, CommandHandler, ContextTypes
import os
from dotenv import load_dotenv

logging.basicConfig(
    level=logging.WARNING,
    format="%(asctime)s [%(levelname)s] %(message)s"
)
logger = logging.getLogger(__name__)

load_dotenv()
TOKEN = os.getenv("TELEGRAM_BOT_TOKEN")

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    msg = "Привет! 👋 Добро пожаловать, как и обещал вот твои файлы :)"
    await update.message.reply_text(msg)
    logger.warning("Message shown to user: %s", msg)

app = ApplicationBuilder().token(TOKEN).build()
app.add_handler(CommandHandler("start", start))

logger.warning("🟢 Telegram бот запущен...")
app.run_polling()
