tg_bot/pyproject.toml
[project]
name = "tg_bot"
version = "0.1.0"
description = "Telegram bot for reading comments and answering to direct"
authors = [{ name = "boxdancer", email = "mr.tselin@yandex.ru" }]
requires-python = ">=3.11"

dependencies = [
    "python-dotenv",
    "python-telegram-bot>=20.0,<21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"